backend/requirements.txt
Django==4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
drf-spectacular

